feat(junit): Add config extension
