+++ utils/junit-utils/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+  `java-library`
+}
+
+apply(from = "$rootDir/gradle/java.gradle")
+
+dependencies {
+  api(libs.bytebuddy)
+  api(libs.bytebuddyagent)
+  api(libs.forbiddenapis)
+  api(project(":components:environment"))
+
+  compileOnly(libs.junit.jupiter)
+}
