+++ utils/test-utils/build.gradle.kts
@@ -11,6 +11,7 @@ dependencies {
 
   api(project(":components:environment"))
   api(project(":utils:config-utils"))
+  api(project(":utils:junit-utils"))
   api(group = "commons-fileupload", name = "commons-fileupload", version = "1.5")
 
   compileOnly(libs.junit.jupiter)
