+++ utils/test-utils/src/main/java/datadog/trace/test/util/DDJavaSpecification.java
@@ -1,39 +1,22 @@
 package datadog.trace.test.util;
 
-import static net.bytebuddy.description.modifier.FieldManifestation.VOLATILE;
-import static net.bytebuddy.description.modifier.Ownership.STATIC;
-import static net.bytebuddy.description.modifier.Visibility.PUBLIC;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.namedOneOf;
-import static net.bytebuddy.matcher.ElementMatchers.none;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import datadog.environment.EnvironmentVariables;
+import datadog.trace.junit.utils.config.WithConfigExtension;
 import de.thetaphi.forbiddenapis.SuppressForbidden;
-import java.lang.instrument.Instrumentation;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
 import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
 import java.util.Set;
 import java.util.stream.Collectors;
-import net.bytebuddy.agent.ByteBuddyAgent;
-import net.bytebuddy.agent.builder.AgentBuilder;
-import net.bytebuddy.dynamic.ClassFileLocator;
-import net.bytebuddy.dynamic.Transformer;
-import net.bytebuddy.utility.JavaModule;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.ExtendWith;
 
+@ExtendWith(WithConfigExtension.class)
 @SuppressForbidden
 public class DDJavaSpecification {
 
@@ -41,22 +24,8 @@ public class DDJavaSpecification {
 
   static final String CONTEXT_BINDER = "datadog.context.ContextBinder";
   static final String CONTEXT_MANAGER = "datadog.context.ContextManager";
-  static final String INST_CONFIG = "datadog.trace.api.InstrumenterConfig";
-  static final String CONFIG = "datadog.trace.api.Config";
-
-  private static Field instConfigInstanceField;
-  private static Constructor<?> instConfigConstructor;
-  private static Field configInstanceField;
-  private static Constructor<?> configConstructor;
 
   private static Boolean contextTestingAllowed;
-  private static volatile boolean isConfigInstanceModifiable = false;
-  static volatile boolean configModificationFailed = false;
-
-  protected static final TestEnvironmentVariables environmentVariables =
-      TestEnvironmentVariables.setup();
-
-  private static Properties originalSystemProperties;
 
   protected boolean assertThreadsEachCleanup = true;
   private static volatile boolean ignoreThreadCleanup;
@@ -64,11 +33,6 @@ public class DDJavaSpecification {
   @BeforeAll
   static void beforeAll() {
     allowContextTesting();
-    installConfigTransformer();
-    makeConfigInstanceModifiable();
-    assertFalse(
-        configModificationFailed,
-        "Config class modification failed. Ensure all test classes extend DDJavaSpecification");
     assertTrue(
         EnvironmentVariables.getAll().entrySet().stream()
             .noneMatch(e -> e.getKey().startsWith("DD_")));
@@ -86,7 +50,6 @@ public class DDJavaSpecification {
           "Found DD threads before test started. Ignoring thread cleanup for this test class");
       ignoreThreadCleanup = true;
     }
-    saveProperties();
   }
 
   static void allowContextTesting() {
@@ -110,87 +73,14 @@ public class DDJavaSpecification {
     }
   }
 
-  private static void installConfigTransformer() {
-    try {
-      Instrumentation instrumentation = ByteBuddyAgent.install();
-      new AgentBuilder.Default()
-          .with(AgentBuilder.RedefinitionStrategy.RETRANSFORMATION)
-          .with(AgentBuilder.RedefinitionStrategy.Listener.ErrorEscalating.FAIL_FAST)
-          .with(
-              new AgentBuilder.LocationStrategy.Simple(
-                  ClassFileLocator.ForClassLoader.ofSystemLoader()))
-          .ignore(none())
-          .type(namedOneOf(INST_CONFIG, CONFIG))
-          .transform(
-              (builder, typeDescription, classLoader, module, pd) ->
-                  builder
-                      .field(named("INSTANCE"))
-                      .transform(Transformer.ForField.withModifiers(PUBLIC, STATIC, VOLATILE)))
-          .with(new ConfigInstrumentationFailedListener())
-          .installOn(instrumentation);
-    } catch (IllegalStateException e) {
-      // Ignore. When we have -javaagent:dd-java-agent.jar, this is fine.
-    }
-  }
-
-  static void makeConfigInstanceModifiable() {
-    if (isConfigInstanceModifiable || configModificationFailed) {
-      return;
-    }
-
-    try {
-      Class<?> instConfigClass = Class.forName(INST_CONFIG);
-      instConfigInstanceField = instConfigClass.getDeclaredField("INSTANCE");
-      instConfigConstructor = instConfigClass.getDeclaredConstructor();
-      instConfigConstructor.setAccessible(true);
-      Class<?> configClass = Class.forName(CONFIG);
-      configInstanceField = configClass.getDeclaredField("INSTANCE");
-      configConstructor = configClass.getDeclaredConstructor();
-      configConstructor.setAccessible(true);
-
-      isConfigInstanceModifiable = true;
-    } catch (ClassNotFoundException e) {
-      if (INST_CONFIG.equals(e.getMessage()) || CONFIG.equals(e.getMessage())) {
-        System.out.println("Config class not found in this classloader. Not transforming it");
-      } else {
-        configModificationFailed = true;
-        System.out.println("Config will not be modifiable");
-        e.printStackTrace();
-      }
-    } catch (ReflectiveOperationException e) {
-      configModificationFailed = true;
-      System.out.println("Config will not be modifiable");
-      e.printStackTrace();
-    }
-  }
-
-  private static void saveProperties() {
-    originalSystemProperties = new Properties();
-    originalSystemProperties.putAll(System.getProperties());
-  }
-
-  private static void restoreProperties() {
-    if (originalSystemProperties != null) {
-      Properties copy = new Properties();
-      copy.putAll(originalSystemProperties);
-      System.setProperties(copy);
-    }
-  }
-
   @AfterAll
   static void afterAll() {
-    restoreProperties();
-
     assertTrue(
         EnvironmentVariables.getAll().entrySet().stream()
             .noneMatch(e -> e.getKey().startsWith("DD_")));
     assertTrue(
         systemPropertiesExceptAllowed().entrySet().stream()
             .noneMatch(e -> e.getKey().toString().startsWith("dd.")));
-
-    if (isConfigInstanceModifiable) {
-      rebuildConfig();
-    }
 
     checkThreads();
   }
@@ -204,39 +94,8 @@ public class DDJavaSpecification {
         .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
   }
 
-  @BeforeEach
-  void setup() {
-    restoreProperties();
-
-    assertTrue(
-        EnvironmentVariables.getAll().entrySet().stream()
-            .noneMatch(e -> e.getKey().startsWith("DD_")));
-    assertTrue(
-        systemPropertiesExceptAllowed().entrySet().stream()
-            .noneMatch(e -> e.getKey().toString().startsWith("dd.")));
-
-    if (isConfigInstanceModifiable) {
-      rebuildConfig();
-    }
-  }
-
   @AfterEach
   void cleanup() {
-    environmentVariables.clear();
-
-    restoreProperties();
-
-    assertTrue(
-        EnvironmentVariables.getAll().entrySet().stream()
-            .noneMatch(e -> e.getKey().startsWith("DD_")));
-    assertTrue(
-        systemPropertiesExceptAllowed().entrySet().stream()
-            .noneMatch(e -> e.getKey().toString().startsWith("dd.")));
-
-    if (isConfigInstanceModifiable) {
-      rebuildConfig();
-    }
-
     if (assertThreadsEachCleanup) {
       checkThreads();
     }
@@ -277,143 +136,4 @@ public class DDJavaSpecification {
     }
   }
 
-  public void injectSysConfig(String name, String value) {
-    injectSysConfig(name, value, true);
-  }
-
-  public void injectSysConfig(String name, String value, boolean addPrefix) {
-    checkConfigTransformation();
-
-    String prefixedName = name.startsWith("dd.") || !addPrefix ? name : "dd." + name;
-    System.setProperty(prefixedName, value);
-    rebuildConfig();
-  }
-
-  public void removeSysConfig(String name) {
-    removeSysConfig(name, true);
-  }
-
-  public void removeSysConfig(String name, boolean addPrefix) {
-    checkConfigTransformation();
-
-    String prefixedName = name.startsWith("dd.") || !addPrefix ? name : "dd." + name;
-    System.clearProperty(prefixedName);
-    rebuildConfig();
-  }
-
-  public void injectEnvConfig(String name, String value) {
-    injectEnvConfig(name, value, true);
-  }
-
-  public void injectEnvConfig(String name, String value, boolean addPrefix) {
-    checkConfigTransformation();
-
-    String prefixedName = name.startsWith("DD_") || !addPrefix ? name : "DD_" + name;
-    environmentVariables.set(prefixedName, value);
-    rebuildConfig();
-  }
-
-  public void removeEnvConfig(String name) {
-    removeEnvConfig(name, true);
-  }
-
-  public void removeEnvConfig(String name, boolean addPrefix) {
-    checkConfigTransformation();
-
-    String prefixedName = name.startsWith("DD_") || !addPrefix ? name : "DD_" + name;
-    environmentVariables.removePrefixed(prefixedName);
-    rebuildConfig();
-  }
-
-  static void rebuildConfig() {
-    synchronized (DDJavaSpecification.class) {
-      checkConfigTransformation();
-      try {
-        Object newInstConfig = instConfigConstructor.newInstance();
-        instConfigInstanceField.set(null, newInstConfig);
-        Object newConfig = configConstructor.newInstance();
-        configInstanceField.set(null, newConfig);
-      } catch (ReflectiveOperationException e) {
-        throw new AssertionError("Failed to rebuild config", e);
-      }
-    }
-  }
-
-  private static void checkConfigTransformation() {
-    assertTrue(isConfigInstanceModifiable);
-    assertNotNull(instConfigConstructor);
-    checkWritable(instConfigInstanceField);
-    assertNotNull(configConstructor);
-    checkWritable(configInstanceField);
-  }
-
-  private static void checkWritable(Field field) {
-    assertNotNull(field);
-    assertTrue(Modifier.isPublic(field.getModifiers()));
-    assertTrue(Modifier.isStatic(field.getModifiers()));
-    assertTrue(Modifier.isVolatile(field.getModifiers()));
-    assertFalse(Modifier.isFinal(field.getModifiers()));
-  }
-
-  public static class TestEnvironmentVariables
-      extends EnvironmentVariables.EnvironmentVariablesProvider {
-    private final Map<String, String> env = new HashMap<>();
-
-    TestEnvironmentVariables(String... kv) {
-      for (int i = 0; i + 1 < kv.length; i += 2) {
-        env.put(kv[i], kv[i + 1]);
-      }
-    }
-
-    @Override
-    public String get(String name) {
-      return env.get(name);
-    }
-
-    @Override
-    public Map<String, String> getAll() {
-      return env;
-    }
-
-    public void set(String name, String value) {
-      env.put(name, value);
-    }
-
-    public void removePrefixed(String prefix) {
-      env.keySet().removeIf(k -> k.startsWith(prefix));
-    }
-
-    public void clear() {
-      env.clear();
-    }
-
-    @SuppressForbidden
-    static TestEnvironmentVariables setup(String... kv) {
-      TestEnvironmentVariables provider = new TestEnvironmentVariables(kv);
-      EnvironmentVariables.provider = provider;
-
-      String propagateVars = System.getenv("TEST_ENV_PROPAGATE_VARS");
-      if (propagateVars != null) {
-        for (String envVar : propagateVars.split(",")) {
-          provider.env.put(envVar, System.getenv(envVar));
-        }
-      }
-
-      return provider;
-    }
-  }
-
-  private static class ConfigInstrumentationFailedListener extends AgentBuilder.Listener.Adapter {
-    @Override
-    public void onError(
-        String typeName,
-        ClassLoader classLoader,
-        JavaModule module,
-        boolean loaded,
-        Throwable throwable) {
-      if (CONFIG.equals(typeName)) {
-        configModificationFailed = true;
-      }
-    }
-  }
 }
